saving before doing redis new user caching

+++ actions/delete-file.ts
@@ -4,7 +4,9 @@ import { UTApi } from "uploadthing/server";
 
 export const DeleteFile = async (value:string) => {
     const utapi = new UTApi()
-    const id = value.split("/")[4]
-    console.log(id)
-    utapi.deleteFiles(id)
+    if (value){
+        const id = value.split("/")[4]
+        console.log(id)
+        utapi.deleteFiles(id)
+    }
 }
